% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{bound_rx}
\alias{bound_rx}
\title{Transform character strings into one regular expression}
\usage{
bound_rx(rx_var, leftbound = "\\\\b", rightbound = "\\\\b", by_element = FALSE)
}
\arguments{
\item{rx_var}{A charater vector}

\item{by_element}{}
}
\value{
An atomic character vector with a valid regex
}
\description{
Passing a list of patterns to a `stringr` function that do not match the length of the `text` argument yields unexpected results.
Converting the expression with a `map` function is cumbersome and sometimes slow.
This transformation solves that problem and may provide an increase in speed 
Using `fixed()` and a `map`function may be faster but will match different things.
}
\examples{

Find all individual letters in a text regardless of case.

sample_text <- c("ag;als", "w a;lsg;lak B", "l A jlajksf k")

str_count(sample_text, bound_rx(letters))
}
