% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{create_boxplot}
\alias{create_boxplot}
\title{Quick plot of a boxplot}
\usage{
create_boxplot(df, x_var, y_var)
}
\arguments{
\item{df}{A data frame or tibble}

\item{x_var}{Normally, a factor variable}

\item{y_var}{Normally, a continuous variable}
}
\value{
A ggplot object.
}
\description{
Quick plot of a boxplot
}
\examples{
create_boxplot(iris, Species, Petal.Width)
}
