% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{create_full_path}
\alias{create_full_path}
\title{Create full paths to files for other functions}
\usage{
create_full_path(path, filename, extension)
}
\arguments{
\item{path}{Path for file as a string with folders created as required.
When \code{NULL}, it defaults to "Processed" under the pwd. Any additional folders
listed here will appear under "Processed". If only one subdirectory required,
it can be a simple string (e.g. "Outputs" will create "Data/Outputs/")
otherwise, a character vector  (c("Outputs", "Maps") for "Data/Outputs/Maps")
If you need to save outside the "Data" folder, use the relevant save function.}
}
\value{
A string with a properly formatted absolute filename from the current working directory
}
\description{
Create full paths to files for other functions
}
\section{Other arguments}{

Passed from relevant save functions
}

