% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_india_map.R
\name{create_india_map}
\alias{create_india_map}
\title{Create maps of India from sf objects with sensible defaults}
\usage{
create_india_map(
  df,
  title = NULL,
  subtitle = NULL,
  fill_var = NULL,
  fill_var_text = NULL,
  cities_w_pop = NULL,
  pp_w_MW = NULL,
  state_labels = "acronym",
  state_labels_colour = "black",
  district_display = "none",
  district_label_colour = "white",
  background_colour = "antiquewhite1",
  legend_omit = FALSE,
  legend_scalar = 2,
  text_scalar = 1,
  legend_fill = "transparent",
  leg_just = c("right", "bottom"),
  leg_x = 0.85,
  leg_y = 0.05,
  show_disputed = FALSE
)
}
\arguments{
\item{df}{A dataframe with a simple features geometry column (representing India of course)}

\item{title}{The title of the plot}

\item{subtitle}{}

\item{fill_var}{The variable to fill the text with. If empty, it takes the 
relevant State variable.}

\item{fill_var_text}{Variable as a string. The labelling to be applied to each observation in df}

\item{cities_w_pop}{Numeric value indicating the minimum population size of
settlements to be plotted. None plotted with default \code{NULL}.}

\item{pp_w_MW}{Numeric value indicating the minimum capacity size of
power plants to be plotted. None plotted with default \code{NULL}.}

\item{state_labels}{One of "none", "type" or "acronym" to indicate 
how states should be labelled. "Type" depicts a letter depending on whether
the state is considered 
Backward (B), Conflictual (C), Med. HDI (M) or High HDI (H).
"Acronym" uses the official abbreviation for
each state.}

\item{state_labels_colour}{Value passed to \code{geom_sf(colour = )}}

\item{district_display}{A choice of whether to have only borders ("borders"), no borders
and no labels ("none"), or both ("labels"). It defaults to "borders".}

\item{district_label_colour}{Value passed to\code{geom_sf_text(colour =)}}

\item{background_colour}{Passed on to \code{theme(panel.background)}}

\item{legend_omit}{Boolean.}

\item{legend_scalar}{To tweak the size of the legend box.}

\item{text_scalar}{To tweak the sizes of most elements.}

\item{legend_fill}{Passed on to \code{theme(legend.background)}}

\item{leg_just}{Passed on to \code{theme(legend.justification)}}

\item{leg_x}{Passed on to \code{theme(legend.position)}}

\item{leg_y}{Passed on to \code{theme(legend.position)}}
}
\value{
A ggplot object
}
\description{
Create maps of India from sf objects with sensible defaults
}
