% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicheck.R
\name{duplicheck}
\alias{duplicheck}
\title{Check for duplicates (in Consumer Pyramids and Indian election data)}
\usage{
duplicheck(df, checkCols = NULL, fromLast = NULL, ...)
}
\arguments{
\item{df}{A dataframe, including sf, potentially containing duplicates.}

\item{checkCols}{Column names as strings that will build the minimum distinct.
dataset to be checked for duplicates. Defaults to ST_NAME/STATE and AC_NO.}

\item{fromLast}{As in \code{duplicated}. The default \code{NULL} means that 
both sides will be checked and returned (i.e. )}

\item{...}{Arguments passed on to \code{duplicated}.}
}
\value{
The same as \code{df} with only the duplicated rows (all of them).
}
\description{
Check for duplicates (in Consumer Pyramids and Indian election data)
}
