% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endwrite.R
\name{endpipe}
\alias{endpipe}
\title{Add a magrittr::pipe at the end of the line and move to the next line}
\usage{
endpipe()
}
\description{
Add a magrittr::pipe at the end of the line and move to the next line
}
\details{
Wrapper around a function that adds code at the end of the currently 
active line and moves the cursor to the next line
}
