% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endwrite.R
\name{endplus}
\alias{endplus}
\title{Add a plus sign ('+') at the end of the line and move to the next line (for ggplot)}
\usage{
endplus()
}
\description{
Add a plus sign ('+') at the end of the line and move to the next line (for ggplot)
}
\details{
Wrapper around a function that adds code at the end of the currently 
active line and moves the cursor to the next line
}
