% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endwrite.R
\name{endwrite}
\alias{endwrite}
\title{Add the relevant sign at the end of the line and move to the next line}
\usage{
endwrite(sign)
}
\description{
Add the relevant sign at the end of the line and move to the next line
}
\details{
Wrapper around a function that adds code at the end of the currently 
active line and moves the cursor to the next line
}
