% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_st_name.R
\name{filter_st_name}
\alias{filter_st_name}
\title{Filtering types of Indian states and/or union territories}
\usage{
filter_st_name(df, states = NA, exclude = TRUE, state_col_name = NULL)
}
\arguments{
\item{df}{An object with class data.frame that has ONE column of Indian states}

\item{states}{What types of top level admin unit should be excluded. The 
default is to filter out those states that are not in Wave 1 
of Consumer Pyramids (but maintaining Union Territories if present). 
Other options are \code{"small_rich"} (Goa, Delhi, Puducherry and Chandigarh), 
\code{"no_vidhan"} (without legislative assembly), and \code{"all_above"}.

It also accepts a character vector of state names.}

\item{exclude}{Do we exclude the matches of state names or do we keep those?}

\item{state_col_name}{If more than column is known to be present, introduce
the name of the correct column here as a string.}
}
\value{
The same object withouth rows referring to the relevant states
}
\description{
Filtering types of Indian states and/or union territories
}
