% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_checks.R
\name{find_seq_incomplete}
\alias{find_seq_incomplete}
\title{Check for gaps in numeric sequence by group}
\usage{
find_seq_incomplete(
  df,
  checkCols = NULL,
  comparison = `==`,
  step = NULL,
  verbose = FALSE
)
}
\arguments{
\item{df}{A dataframe}

\item{checkCols}{A character vector with exactly two strings. 
The first names the grouping variable, and the second,
the accompanying _numeric_ sequence variable.}

\item{comparison}{A relational operator for the check, with back-ticks.}

\item{step}{Step to be used in comparison. Defaults to the difference 
between the first and second elements of the vector.}

\item{verbose}{Should messages be reported?}
}
\value{
If any gaps found, the function returns a data frame with two rows 
before and one after the gap in the sequence, extracted by group. 
If no gaps found, NULL invisibly.
}
\description{
Identifies **gaps** only, if the sequence is broken
because of a duplication, use \code{duplicheck}
}
