% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{find_st_name}
\alias{find_st_name}
\title{Find the element containing state names}
\usage{
find_st_name(obj, df_nm = "<your_object>")
}
\arguments{
\item{df}{A dataframe with one or more columns containing Indian state names}
}
\value{
The name of the column as a string literal.
}
\description{
Find the element containing state names
}
\details{
The function first checks if "ST_NAME" or "STATE" are present
preferring always one of those two, and always in that order. Otherwise,
it will search in character columns for those containing the word "Pradesh"
(Hindi for "state" or "province").
}
