% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_seq_complete.R
\name{is.sequential}
\alias{is.sequential}
\title{Check whether there are different steps between the elements of a vector}
\usage{
is.sequential(x, comparison, step)
}
\arguments{
\item{x}{A numeric vector.}

\item{comparison}{A relational operator for the check, with back-ticks.}

\item{step}{Step to be used in comparison. Defaults to the difference 
between the first and second elements of the vector.}
}
\value{
A logical vector
}
\description{
Check whether there are different steps between the elements of a vector
}
