% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_seq_complete.R
\name{is_acno_complete}
\alias{is_acno_complete}
\title{Check for gaps in AC No. sequence in Indian states}
\usage{
is_acno_complete(df, checkCols = NULL, comparison = `==`, step = NULL)
}
\arguments{
\item{df}{A dataframe. The default assumes one column for state names 
and one column for
AC numbers. The state column name is automatically detected and
the AC number column name assumed to be either AC_NO or ac_no.}

\item{checkCols}{To override defaults, provide a character vector with 
exactly two strings. The first names the grouping variable, and the second,
the accompanying numeric sequence variable.}

\item{comparison}{A relational operator for the check, with back-ticks.}

\item{step}{Step to be used in comparison. Defaults to the difference 
between the first and second elements of the vector.}
}
\value{
A data frame with two AC_NO before and one after the fault 
in the AC_NO sequence, extracted per state.
}
\description{
Identifies **gaps** only, if the sequence is broken
because of a duplication, use \code{duplicheck}
}
