% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_seq_complete.R
\name{is_seq_complete}
\alias{is_seq_complete}
\title{Check for gaps in numeric sequence by group}
\usage{
is_seq_complete(df, checkCols = NULL)
}
\arguments{
\item{df}{A dataframe. The default assumes one column for state names 
and one column for
AC numbers. The state column name is automatically detected and
the AC number column name assumed to be either AC_NO or ac_no.}

\item{checkCols}{To override defaults, provide a character vector with 
exactly two strings. The first names the grouping variable, and the second,
the accompanying numeric sequence variable.}
}
\value{
A data frame with two rows before and one after the gap 
in the sequence, extracted by group.
}
\description{
wrapper around \code{is_acno_complete}, which was
originally created for checking which Assembly Constituency numbers 
are missing 
in a dataframe with Indian election data. 

Identifies **gaps** only, if the sequence is broken
because of a duplication, use \code{duplicheck}.
}
