% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_PDF.R
\name{map_PDF}
\alias{map_PDF}
\title{Saving multiple plots into PDF from a list}
\usage{
map_PDF(plot_list, multiple = FALSE, nrow = 1, ncol = 1)
}
\arguments{
\item{plot_list}{A list of plots handled by gridExtra}

\item{multiple}{Should the images be printed in multiple PDF files?}

\item{nrow}{Number of rows for gridExtra::marrangeGrob}

\item{ncol}{Number of columns for gridExtra::marrangeGrob}
}
\value{
One or several PDF file with numnber of map per page or one PDF per map 
in separate folder
}
\description{
Saving multiple plots into PDF from a list
}
