% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{match_col_classes}
\alias{match_col_classes}
\title{Match column classes of df before joining}
\usage{
match_col_classes(df1, df2)
}
\arguments{
\item{df1}{}

\item{df2}{}
}
\value{
df2 with its shared columns (the ones with the same name) in the same class as df1
}
\description{
Match column classes of df before joining
}
