% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mismatch.R
\name{mismatch}
\alias{mismatch}
\title{Search for Elements in Vector2 Missing from vector2 and Vice Versa}
\usage{
mismatch(
  vector1,
  vector2,
  min_obs_for_csv = NULL,
  folder = "mismatch",
  treatment = as.character
)
}
\arguments{
\item{vector1}{A vector for comparison with \%in\%. Geometry columns will be
removed if vector originates in sf object.}

\item{vector2}{As for vector1}

\item{min_obs_for_csv}{If NULL, output only to console. An integer value
indicates the minimum number of mismatched observations for the output to
be stored in a file in the folder given below.}

\item{treatment}{The coercing method for the vectors}
}
\value{
A tibble with the mismatches, including missing values.
}
\description{
This function is intended to offer a quick visualisation of
  which elements are missing across two vectors.
}
\examples{
mismatch(iris$Species, iris$Species)

# Currently gives one NA value if all correct.
}
