% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise}
\alias{normalise}
\title{Normalise values in a vector (default 0-1)}
\usage{
normalise(v, min_val, max_val)
}
\arguments{
\item{v}{A numeric vector}

\item{min_val}{A minimum value}

\item{max_val}{A maximum value}
}
\value{
A numeric vector
}
\description{
Normalise values in a vector (default 0-1)
}
