% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odd_val_col.R
\name{odd_val_col}
\alias{odd_val_col}
\title{Checking Columns for NA, Negative or Other (Odd) Values and Returning an
easy to inspect dataframe with meaningful reference columns that don't have
NAs}
\usage{
odd_val_col(
  df,
  refCols = c(1, 2),
  oddValues = "NAs",
  states = "not_in_cpw1",
  verbose = TRUE
)
}
\arguments{
\item{df}{Dataframe in which to check for odd values.}

\item{refCols}{The columns that will be presented in the output. Passed to a 
If \code{"dist"} or \code{"ac"} are typed in, the function selects the pairs 
STATE/DISTRICT_NAME and ST_NAME/AC_NO.}

\item{oddValues}{The type of odd values with three options. The default
\code{"na"} checks all types of missing values in Consumer Pyramids. 
The option \code{"negative"} checks if any
numeric column is negative. Alternatively, a string vector can be supplied
that will used to check.}

\item{states}{What types of top level admin unit should be excluded. The 
default is to filter out those states that are not in Wave 1 
of Consumer Pyramids (but maintaining Union Territories if present). 
Other options are \code{"small_rich"} (Goa, Delhi, Puducherry and Chandigarh), 
\code{"no_vidhan"} (without legislative assembly), and \code{"all_above"}.

It also accepts a character vector of state names.}

\item{verbose}{Whether to print messages to the console.}

\item{df_type}{as \code{refCols}.}
}
\value{
A tibble with the rows where the odd values where found.
}
\description{
Checking Columns for NA, Negative or Other (Odd) Values and Returning an
easy to inspect dataframe with meaningful reference columns that don't have
NAs
}
\examples{

x <- na_if(mtcars, 0)

odd_val_col(x)
}
