% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{peek_names}
\alias{peek_names}
\title{Checking if object contains a pattern in its names}
\usage{
peek_names(obj, pattern, ...)
}
\arguments{
\item{obj}{An R object with names}

\item{pattern}{A regex pattern as a character vector to use with grepl}

\item{...}{Additional arguments passed to grepl}
}
\value{
the matches of `pattern` in `obj` names
}
\description{
Checking if object contains a pattern in its names
}
\examples{

peek_names(mtcars)

}
