% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{rcsv}
\alias{rcsv}
\title{Read a CSV File and Create Object with Same Name}
\usage{
rcsv(object, path = "Processed/", verbose = FALSE)
}
\arguments{
\item{object}{Filename, which will become object name}

\item{path}{Directory to look for filename}
}
\value{
A tibble
}
\description{
Read a CSV File and Create Object with Same Name
}
