% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rename_geometry}
\alias{rename_geometry}
\title{Renames an sf object geometry column (e.g. to allow rbind to work)}
\usage{
rename_geometry(obj, name)
}
\arguments{
\item{obj}{An sf object}

\item{name}{The new name}
}
\value{
An sf object
}
\description{
Renames an sf object geometry column (e.g. to allow rbind to work)
}
