% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rename_states}
\alias{rename_states}
\title{Rename states}
\usage{
rename_states(df)
}
\arguments{
\item{df}{dataframe with a column containing Indian states (automatically found by \code{find_st_name_col}), potentially with 
older names like 'Orissa', 'Pondicherry', etc.}
}
\value{
The same dataframe but with STATE/ST_NAME columm recoded
}
\description{
Rename states
}
