% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{replace_xy}
\alias{replace_xy}
\title{Replace NA in .x  cols by .y cols (or vice versa) after incomplete *_join}
\usage{
replace_xy(tbl, vars = NULL, ending_to_keep = ".x")
}
\arguments{
\item{tbl}{A tibble or data_frame}

\item{vars}{By default, the function searches for variables with .x/.y suffixes.
Alternatively, a character vector of names of variables that are duplicated
can be provided. This is useful when operating with \code{sf} objects because 'geometry'
column will throw an error with the automated search.
These variable must be entered without .x or .y suffixes.}

\item{ending_to_keep}{Choose which column suffix should be kept for rows
where both columns have values. Currently only implemented for .x/.y.}
}
\value{
A tibble with merged columns
}
\description{
Replace NA in .x  cols by .y cols (or vice versa) after incomplete *_join
}
