% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rm_list_cols}
\alias{rm_list_cols}
\title{Remove list columns}
\usage{
rm_list_cols(x)
}
\arguments{
\item{x}{An R object with the class "data.frame"; others will be ignored and
the original will be silently returned untouched.}
}
\value{
The same object as x but without the list columns
}
\description{
Remove list columns
}
