% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ins.R
\name{set_new_chapter}
\alias{set_new_chapter}
\title{Insert dashes from cursor position to up to 80 characters}
\usage{
set_new_chapter()
}
\value{
dashes inside RStudio
}
\description{
Insert dashes from cursor position to up to 80 characters
}
