% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{sf_GPKG}
\alias{sf_GPKG}
\title{Quick overwrite GPKG file}
\usage{
sf_GPKG(x, path = NULL, overwrite = TRUE)
}
\arguments{
\item{x}{An sf object}

\item{path}{Path for file as a string with folders created as required.
When \code{NULL}, it defaults to "Processed" under the pwd. Any additional folders
listed here will appear under "Processed". If only one subdirectory required,
it can be a simple string (e.g. "Outputs" will create "Data/Outputs/")
otherwise, a character vector  (c("Outputs", "Maps") for "Data/Outputs/Maps")
If you need to save outside the "Data" folder, use the relevant save function.}

\item{overwrite}{The purpose of this function is to facilitate overwriting GPKG files.
However, in some situations, it may be best to change this switch to FALSE for a prompt.
If overwrite is rarely the desired outcome, use the default sf::st_write function.}
}
\value{
A GPKG file with the same name as \code{x}
}
\description{
Quick overwrite GPKG file
}
