% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sfplot}
\alias{sfplot}
\title{Quick plot of an sf object}
\usage{
sfplot(df, colour = "red", fill = NA, label_col_int = FALSE)
}
\arguments{
\item{df}{An sf object.}

\item{colour}{Colour of borders.}

\item{fill}{Main variable in \code{df} to be represented in plot, as a string.
Defaults to grey.}

\item{label_col_int}{Integer value of column containing labels}
}
\value{
A ggplot object.
}
\description{
Quick plot of an sf object
}
\examples{
\dontrun{sfplot(CP_dist_map, fill = "MEDIAN_HRS_ACCESS", label_col_int = 2)}
}
