% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{sfread}
\alias{sfread}
\title{Read a GPKG File and Create Object with Same Name}
\usage{
sfread(object, path = "Processed/", verbose = FALSE)
}
\arguments{
\item{object}{Filename, which will become object name}

\item{path}{Directory to look for filename}

\item{verbose}{Boolean. Display loading messages?}
}
\value{
An sf object
}
\description{
Read a GPKG File and Create Object with Same Name
}
