% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{View}
\alias{View}
\title{View without list-columns or geometry}
\usage{
View(x, title = NULL)
}
\arguments{
\item{x}{An R object}

\item{title}{The title of the (not working in tibble::utils)}
}
\value{
Both the object (now without list-columns) and its visualisation
}
\description{
R's View() always tries to show geometry and other list columns
but, most of the time, this is rather useless and slows things down.
This function is designed to mask the default View(). It removes any list-columns,
including geometry columns from data.frame objects.
Using code modified from tibble::view(), it also allows RStudio to display
and returns the object invisibly.
}
\examples{
iris_same <- iris \%>\%
  mutate(listcol = purrr::map(Petal.Width, ~`+`(Petal.Length, .x))) \%>\%
  as_tibble() \%>\%
  View()

iris_same
}
