% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{wcsv}
\alias{wcsv}
\title{Quick save CSV file}
\usage{
wcsv(df, path = NULL)
}
\arguments{
\item{df}{A tibble}

\item{path}{Path for file as a string with folders created as required.
When \code{NULL}, it defaults to "Processed" under the pwd. Any additional folders
listed here will appear under "Processed". If only one subdirectory required,
it can be a simple string (e.g. "Outputs" will create "Data/Outputs/")
otherwise, a character vector  (c("Outputs", "Maps") for "Data/Outputs/Maps")
If you need to save outside the "Data" folder, use the relevant save function.}
}
\value{
A CSV file with the same name as \code{df}
}
\description{
Quick save CSV file
}
