% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{date_str}
\alias{date_str}
\title{Return a filename-friendly string with the current time with the typical defaults}
\usage{
date_str(tzone = "Australia/Melbourne", format = "\%Y\%m\%d_\%H\%M")
}
\arguments{
\item{tzone}{OlsonName compliant string. Passed as "tzone" attribute to POSIXct object}

\item{format}{sprintf compliant. Passed to \code{format}.}
}
\value{

}
\description{
Return a filename-friendly string with the current time with the typical defaults
}
\examples{

paste0("/my/path/file_from_date_", date_str())

}
