% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_st_name.R
\name{exclude_states}
\alias{exclude_states}
\title{Exclude (types of) Indian states or Union Territories}
\usage{
exclude_states(df, states, state_col_name = NULL)
}
\arguments{
\item{df}{An object with class data.frame that has ONE column of Indian states.}

\item{states}{Accepts either a character vector of state names, which will be made into a regex with \code{bound_rx(vector, "", "")} and ignore case, or the following options: 
\itemize{
 \item{\code{"not_in_cpw1"} for those not in Consumer Pyramids Wave 1}
 \item{\code{"small_rich"}: Goa, Delhi, Puducherry and Chandigarh}
 \item{\code{"no_vidhan"}: those Union Territories without legislative assembly}
 \item{\code{"northeast"}: Assam, Arunachal Pradesh, Nagaland, Manipur, Mizoram, Tripura, Meghalaya, Sikkim)}
 \item{\code{"all_above"}: all states in previous options}
}}

\item{state_col_name}{If more than column is known to be present, introduce
the name of the correct column here as a string.}
}
\value{
The same object WITHOUT rows referring to the relevant states

#'

The same data.frame object WITHOUT rows referring to the relevant states
}
\description{
Exclude (types of) Indian states or Union Territories
}
