% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_st_name.R
\name{exclude_states}
\alias{exclude_states}
\title{Exclude (types of) Indian states or Union Territories}
\usage{
exclude_states(
  df,
  ...,
  state_col_name = NULL,
  fixed = FALSE,
  ignore_case = TRUE
)
}
\arguments{
\item{df}{An object with class data.frame that has a column of Indian states.}

\item{...}{State or Union Territory names either in full or in part. Case
insensitive. Accepts symbols (bare text) or quoted strings (see details
below). If a symbol refers to a character vector in the global environment,
the vector contents and not the symbol will be part of the search. The
following options are special and can be used in addition to others both as
strings and as symbols: \itemize{ \item{\code{"poor_coverage"}: Jammu &
Kashmir and Uttarakhand because of frequent problems in maps and dataset
coverage} \item{\code{"not_in_cpw1"} for those not in Consumer Pyramids
Wave 1} \item{\code{"small_rich"}: Goa, Delhi, Puducherry and Chandigarh}
\item{\code{"no_vidhan"}: those Union Territories without legislative
assembly} \item{\code{"northeast"}: Assam, Arunachal Pradesh, Nagaland,
Manipur, Mizoram, Tripura, Meghalaya, Sikkim)} \item{\code{"all_above"}:
all states in previous options} }}

\item{state_col_name}{If more than column is known to be present, introduce
the name of the correct column here as a string.}

\item{fixed}{should the names be matched strictly.}

\item{ignore_case}{should case be ignored.}
}
\value{
The same object WITHOUT rows referring to the relevant states
}
\description{
Exclude (types of) Indian states or Union Territories
}
\details{
The elements from ... are processed and coerced into a string vector
that is then turned into an atomic regex by \code{cptools::bound_rx(vector,
  "", "")}, and finally passed to \code{grepl}.

  \code{exclude_states(df, and, "Mizoram", my_states)} will exclude Andhra
  Pradesh, Mizoram and the contents of \code{my_states} if it is in the Global
  Environment
}
