% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mismatch.R
\name{mismatch}
\alias{mismatch}
\title{Search for Elements in Vector2 Missing from vector2 and Vice Versa}
\usage{
mismatch(
  vector1,
  vector2,
  min_obs_for_csv = 20,
  treatment = as.character,
  store_file = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{vector1}{A vector for comparison with \%in\%. Geometry columns will be removed if vector originates in sf object.}

\item{vector2}{As for vector1}

\item{min_obs_for_csv}{If no. of mismatches is above this number,
a CSV will be created under "Processed/Checks/mm_" (directory not created)}

\item{treatment}{The coercing method for the vectors}

\item{store_file}{Should the mismatches be stored separately}

\item{verbose}{Should messages be reported?}
}
\value{
A tibble with the mismatches, including missing values.
}
\description{
This function is intended to offer a quick visualisation of which districts or others are missing in Consumer Pyramids data.
}
\examples{
mismatch(iris$Species, iris$Species)

# Currently gives one NA value if all correct.
}
