% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odd_val_col.R
\name{odd_val_col}
\alias{odd_val_col}
\title{Checking Columns for NA, Negative or Other (Odd) Values and Returning an
easy to inspect dataframe with meaningful reference columns that don't have
NAs}
\usage{
odd_val_col(
  df,
  refCols = c(1, 2),
  oddValues = "NAs",
  states = NA,
  verbose = TRUE
)
}
\arguments{
\item{df}{Dataframe in which to check for odd values.}

\item{refCols}{The columns that will be presented in the output. Passed to a 
If \code{"dist"} or \code{"ac"} are typed in, the function selects the pairs 
STATE/DISTRICT_NAME and ST_NAME/AC_NO.}

\item{oddValues}{The type of odd values with three options. The default
\code{"na"} checks all types of missing values in Consumer Pyramids. 
The option \code{"negative"} checks if any
numeric column is negative. Alternatively, a string vector can be supplied
that will used to check.}

\item{states}{Accepts either a character vector of state names, which will be made into a regex with \code{bound_rx(vector, "", "")} and ignore case, or the following options: 
\itemize{
\item{\code{"poor_coverage"}: Jammu & Kashmir and Uttarakhand because of frequent problems in maps and dataset coverage}
 \item{\code{"not_in_cpw1"} for those not in Consumer Pyramids Wave 1}
 \item{\code{"small_rich"}: Goa, Delhi, Puducherry and Chandigarh}
 \item{\code{"no_vidhan"}: those Union Territories without legislative assembly}
 \item{\code{"northeast"}: Assam, Arunachal Pradesh, Nagaland, Manipur, Mizoram, Tripura, Meghalaya, Sikkim)}
 \item{\code{"all_above"}: all states in previous options}
}}

\item{verbose}{Whether to print messages to the console.}

\item{df_type}{as \code{refCols}.}
}
\value{
A tibble with the rows where the odd values where found.
}
\description{
Checking Columns for NA, Negative or Other (Odd) Values and Returning an
easy to inspect dataframe with meaningful reference columns that don't have
NAs
}
\details{
The \code{states} argument calls \code{exclude_states}
}
\examples{

x <- na_if(mtcars, 0)

odd_val_col(x)
}
