% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/View.R
\name{View}
\alias{View}
\title{View (without crashing Rstudio)}
\usage{
View(x, title = NULL, block_large_df = TRUE, rm_lc = TRUE)
}
\arguments{
\item{x}{An R object}

\item{title}{An optional title chosen for the window that will display the object}

\item{block_large_df}{Should large dataframes be blocked from displaying?}

\item{rm_lc}{Should list-columns be removed?}
}
\value{
Both the object (invisibly) and its display
}
\description{
R's View() always tries to open any object; but some common objects make Rstudio crash or slow things down.
This function is designed to mask the default View(). This version removes by default any list-columns,
including geometry columns from \code{data.frame} objects. It also blocks viewing large dataframes.
}
\details{
Global options can be set to alter behaviour:
\itemize{
 \item{"cpview_mblim":} {an integer for maximum dataframe size in Mb (rough approximation)}
 \item{"cpview_celllim":} {an integer for maximum number of dataframe columns}
 \item{"cpview_warn":} {anything but `NULL` to silence warnings when removing list-columns}
}
}
\examples{
iris_same <- iris \%>\%
  mutate(listcol = purrr::map(Petal.Width, ~`+`(Petal.Length, .x))) \%>\%
  as_tibble() \%>\%
  View()

iris_same
}
